refactor: reorganize backend routes by consolidating user and assessment routes into structured modules, enhancing clarity and maintainability
