+++ backend/routes/setup/database/crud.js
@@ -1,72 +1,14 @@
 import express from 'express';
-import sql from 'mssql';
-import db from '../db/index.js';
+import db from '../../../db/index.js';
 
 const router = express.Router();
 
-// A simple endpoint to test the serverless function
-router.get('/', (req, res) => {
-  res.json({ 
-    message: 'Serverless function is running',
-    environment: process.env.NODE_ENV,
-    timestamp: new Date().toISOString()
-  });
-});
-
-// Test Azure SQL connection
-router.get('/azure-sql-test', async (req, res) => {
-  try {
-    // Determine if we're using Azure SQL
-    const isAzureSql = db.client.config.client === 'mssql';
-    
-    if (!isAzureSql) {
-      console.log('Warning: Not using Azure SQL, falling back to SQLite');
-    }
-    
-    // Test a simple query
-    const result = await db.raw('SELECT 1 as testValue');
-    
-    let testValue;
-    if (isAzureSql) {
-      testValue = result[0].testValue;
-    } else {
-      testValue = result[0].testValue;
-    }
-    
-    res.json({
-      success: true,
-      message: `Successfully connected to ${isAzureSql ? 'Azure SQL' : 'SQLite database'}`,
-      data: {
-        currentTime: new Date().toISOString(),
-        testValue: testValue,
-        databaseType: db.client.config.client,
-        serverInfo: {
-          name: isAzureSql ? db.client.config.connection.server : 'local SQLite',
-          database: isAzureSql ? db.client.config.connection.database : 'SQLite database',
-        }
-      }
-    });
-  } catch (error) {
-    console.error('Database connection test error:', error);
-    res.status(500).json({
-      success: false,
-      message: 'Failed to connect to database',
-      error: {
-        name: error.name,
-        message: error.message,
-      }
-    });
-  }
-});
-
-// Test with a more comprehensive DB operation
-router.get('/azure-sql-crud-test', async (req, res) => {
+router.get("/crud-test", async (req, res) => {
   try {
     const isAzureSql = db.client.config.client === 'mssql';
-    const tableName = 'VercelTest' + Date.now(); // Use a unique table name
+    const tableName = 'VercelTest' + Date.now();
     
     if (isAzureSql) {
-      // Create a test table if it doesn't exist
       await db.raw(`
         IF NOT EXISTS (SELECT * FROM sys.tables WHERE name = '${tableName}')
         BEGIN
@@ -78,7 +20,6 @@ router.get('/azure-sql-crud-test', async (req, res) => {
         END
       `);
     } else {
-      // Create a test table in SQLite
       await db.schema.createTable(tableName, (table) => {
         table.increments('id');
         table.string('message');
@@ -86,7 +27,6 @@ router.get('/azure-sql-crud-test', async (req, res) => {
       });
     }
     
-    // Insert a test record
     let newId;
     const testMessage = 'Test from Vercel ' + new Date().toISOString();
     
@@ -102,7 +42,6 @@ router.get('/azure-sql-crud-test', async (req, res) => {
       newId = insertedIds[0];
     }
     
-    // Read the record back
     let readResult;
     if (isAzureSql) {
       const result = await db.raw(`SELECT * FROM ${tableName} WHERE id = ${newId}`);
@@ -111,7 +50,6 @@ router.get('/azure-sql-crud-test', async (req, res) => {
       readResult = await db(tableName).where('id', newId).first();
     }
     
-    // Get total count
     let totalCount;
     if (isAzureSql) {
       const result = await db.raw(`SELECT COUNT(*) as count FROM ${tableName}`);
@@ -121,7 +59,6 @@ router.get('/azure-sql-crud-test', async (req, res) => {
       totalCount = result[0].count;
     }
     
-    // Clean up - drop the table
     if (isAzureSql) {
       await db.raw(`DROP TABLE ${tableName}`);
     } else {
