+++ backend/routes/setup/database/status.js
@@ -3,7 +3,7 @@ import db from '../../../db/index.js';
 
 const router = express.Router();
 
-router.get("/api/db-status", async (req, res) => {
+router.get("/status", async (req, res) => {
   try {
     // Try to query the database
     await db.raw("SELECT 1");
