+++ backend/routes/setup/health/hello.js
@@ -2,7 +2,7 @@ import express from 'express';
 
 const router = express.Router();
 
-router.get("/api/hello", (req, res) => {
+router.get("/hello", (req, res) => {
   res.json({ message: "Hello World from Dottie API!" });
 });
