+++ backend/routes/setup/middleware/index.js
@@ -1,0 +1,8 @@
+import express from 'express';
+import errorHandlers from './errorHandlers.js';
+
+const router = express.Router();
+
+router.use(errorHandlers);
+
+export default router;
