--- backend/routes/userRoutes.js
@@ -1,91 +1,0 @@
-import express from 'express';
-import User from '../models/User.js';
-
-const router = express.Router();
-
-// Get all users
-router.get('/', async (req, res) => {
-  try {
-    const users = await User.getAll();
-    res.json(users);
-  } catch (error) {
-    console.error('Error fetching users:', error);
-    res.status(500).json({ error: 'Failed to fetch users' });
-  }
-});
-
-// Get user by ID
-router.get('/:id', async (req, res) => {
-  try {
-    const user = await User.findById(req.params.id);
-    if (!user) {
-      return res.status(404).json({ error: 'User not found' });
-    }
-    res.json(user);
-  } catch (error) {
-    console.error('Error fetching user:', error);
-    res.status(500).json({ error: 'Failed to fetch user' });
-  }
-});
-
-// Create a new user
-router.post('/', async (req, res) => {
-  try {
-    const { username, email, password_hash, age } = req.body;
-    
-    // Simple validation
-    if (!username || !email || !password_hash) {
-      return res.status(400).json({ error: 'Username, email, and password_hash are required' });
-    }
-    
-    // Check if user already exists
-    const existingUser = await User.findByEmail(email);
-    if (existingUser) {
-      return res.status(409).json({ error: 'User with this email already exists' });
-    }
-    
-    const user = await User.create({ username, email, password_hash, age });
-    res.status(201).json(user);
-  } catch (error) {
-    console.error('Error creating user:', error);
-    res.status(500).json({ error: 'Failed to create user' });
-  }
-});
-
-// Update a user
-router.put('/:id', async (req, res) => {
-  try {
-    const { username, email, age } = req.body;
-    
-    // Check if user exists
-    const user = await User.findById(req.params.id);
-    if (!user) {
-      return res.status(404).json({ error: 'User not found' });
-    }
-    
-    const updatedUser = await User.update(req.params.id, { username, email, age });
-    res.json(updatedUser);
-  } catch (error) {
-    console.error('Error updating user:', error);
-    res.status(500).json({ error: 'Failed to update user' });
-  }
-});
-
-// Delete a user
-router.delete('/:id', async (req, res) => {
-  try {
-    // Check if user exists
-    const user = await User.findById(req.params.id);
-    if (!user) {
-      return res.status(404).json({ error: 'User not found' });
-    }
-    
-    await User.delete(req.params.id);
-    res.status(204).send();
-  } catch (error) {
-    console.error('Error deleting user:', error);
-    res.status(500).json({ error: 'Failed to delete user' });
-  }
-});
-
-export default router;
